+++ config_controller_primary.sh
@@ -25,7 +25,11 @@ ANSIBLE_PLAYBOOKS="%%ANSIBLE_PLAYBOOKS%%"
 
 INTERFACES="/etc/network/interfaces"
 INTERFACES_D="/etc/network/interfaces.d"
+SWIFT_ENABLED=0
 
+if [ "$ANSIBLE_PLAYBOOKS" = "all+swift" ] || [ "$ANSIBLE_PLAYBOOKS" = "minimal+swift" ]; then
+  SWIFT_ENABLED=1
+fi
 
 apt-get update
 apt-get install -y python-dev python-pip bridge-utils git lvm2 vim xfsprogs
@@ -166,16 +170,18 @@ EOF
 
 ifup -a
 
-pvcreate /dev/xvde1
-vgcreate swift /dev/xvde1
+if [ $SWIFT_ENABLED -eq 1 ]; then
+  pvcreate /dev/xvde1
+  vgcreate swift /dev/xvde1
 
-for DISK in disk1 disk2 disk3; do
-  lvcreate -L 10G -n ${DISK} swift
-  echo "/dev/swift/${DISK} /srv/${DISK} xfs loop,noatime,nodiratime,nobarrier,logbufs=8 0 0" >> /etc/fstab
-  mkfs.xfs -f /dev/swift/${DISK}
-  mkdir -p /srv/${DISK}
-  mount /srv/${DISK}
-done
+  for DISK in disk1 disk2 disk3; do
+    lvcreate -L 10G -n ${DISK} swift
+    echo "/dev/swift/${DISK} /srv/${DISK} xfs loop,noatime,nodiratime,nobarrier,logbufs=8 0 0" >> /etc/fstab
+    mkfs.xfs -f /dev/swift/${DISK}
+    mkdir -p /srv/${DISK}
+    mount /srv/${DISK}
+  done
+fi
 
 rpc_user_config="/etc/rpc_deploy/rpc_user_config.yml"
 swift_config="/etc/rpc_deploy/conf.d/swift.yml"
@@ -184,7 +190,6 @@ user_variables="/etc/rpc_deploy/user_variables.yml"
 echo -n "%%PRIVATE_KEY%%" > .ssh/id_rsa
 chmod 600 .ssh/*
 
-cd /root
 git clone -b %%RPC_GIT_VERSION%% %%RPC_GIT_REPO%% ansible-lxc-rpc
 cd ansible-lxc-rpc
 pip install -r requirements.txt
@@ -200,7 +205,7 @@ sed -i "s/\(rackspace_cloud_password\): .*/\1: %%RACKSPACE_CLOUD_PASSWORD%%/g" $
 sed -i "s/\(rackspace_cloud_api_key\): .*/\1: %%RACKSPACE_CLOUD_API_KEY%%/g" $user_variables
 sed -i "s/\(glance_default_store\): .*/\1: %%GLANCE_DEFAULT_STORE%%/g" $user_variables
 
-if [ "$ANSIBLE_PLAYBOOKS" = "all+swift" ] || [ "$ANSIBLE_PLAYBOOKS" = "minimal+swift" ]; then
+if [ $SWIFT_ENABLED -eq 1 ]; then
   sed -i "s/\(glance_swift_store_auth_address\): .*/\1: '{{ auth_identity_uri }}'/" $user_variables
   sed -i "s/\(glance_swift_store_key\): .*/\1: '{{ glance_service_password }}'/" $user_variables
   sed -i "s/\(glance_swift_store_region\): .*/\1: RegionOne/" $user_variables
@@ -213,13 +218,15 @@ environment_version=$(md5sum /etc/rpc_deploy/rpc_environment.yml | awk '{print $
 
 RAW_URL=$(echo %%HEAT_GIT_REPO%% | sed -e 's/github.com/raw.githubusercontent.com/g')
 
-curl -o $rpc_user_config ${RAW_URL}/%%HEAT_GIT_VERSION%%/rpc_user_config.yml
+curl -o $rpc_user_config "${RAW_URL}/%%HEAT_GIT_VERSION%%/rpc_user_config.yml"
 sed -i "s/__ENVIRONMENT_VERSION__/$environment_version/g" $rpc_user_config
 sed -i "s/__EXTERNAL_VIP_IP__/%%EXTERNAL_VIP_IP%%/g" $rpc_user_config
 sed -i "s/__CLUSTER_PREFIX__/%%CLUSTER_PREFIX%%/g" $rpc_user_config
 
-curl -o $swift_config ${RAW_URL}/%%HEAT_GIT_VERSION%%/swift.yml
-sed -i "s/__CLUSTER_PREFIX__/%%CLUSTER_PREFIX%%/g" $swift_config
+if [ $SWIFT_ENABLED -eq 1 ]; then
+  curl -o $swift_config "${RAW_URL}/%%HEAT_GIT_VERSION%%/swift.yml"
+  sed -i "s/__CLUSTER_PREFIX__/%%CLUSTER_PREFIX%%/g" $swift_config
+fi
 
 cd rpc_deployment
 retry 3 ansible-playbook -e @${user_variables} playbooks/setup/host-setup.yml
@@ -239,7 +246,7 @@ if [ "$ANSIBLE_PLAYBOOKS" = "minimal" ] || [ "$ANSIBLE_PLAYBOOKS" = "minimal+swi
   retry 3 ansible-playbook -e @${user_variables} playbooks/openstack/openstack-setup-no-logging.yml
 fi
 
-if [ "$ANSIBLE_PLAYBOOKS" = "all+swift" ] || [ "$ANSIBLE_PLAYBOOKS" = "minimal+swift" ]; then
+if [ $SWIFT_ENABLED -eq 1 ]; then
   retry 3 ansible-playbook -e @${user_variables} playbooks/openstack/swift-all.yml
 fi
 %%CURL_CLI%% --data-binary '{"status": "SUCCESS"}'
